// wishlist-fe/src/app/enums/product.enum.ts
export enum ProductCategory {
  SOFTWARE = "SOFTWARE",
  CLOTHES = "CLOTHES",
  DEVICE = "DEVICE",
}

export enum ProductType {
  GAME = "GAME",
  WORK = "WORK",
  STUDY = "STUDY",
  HAT = "HAT",
  SHIRT = "SHIRT",
  PANT = "PANT",
  SHOE = "SHOE",
  SMARTPHONE = "SMARTPHONE",
  LAPTOP = "LAPTOP",
  TABLET = "TABLET",
}